pathway-backend/requirements.txt
pathway
fastapi
uvicorn
python-dotenv
google-generativeai

requests
